+++ conan/api/subapi/download.py
@@ -1,7 +1,7 @@
 from conan.api.model import Remote
 from conan.api.subapi import api_method
 from conan.api.output import ConanOutput
-from conan.api.conan_app import ConanApp
+from conan.internal.conan_app import ConanApp
 from conans.client.source import retrieve_exports_sources
 from conans.errors import ConanException
 from conans.model.package_ref import PkgReference
