+++ conan/api/subapi/remove.py
@@ -1,6 +1,6 @@
 from conan.api.model import Remote
 from conan.api.subapi import api_method
-from conan.api.conan_app import ConanApp
+from conan.internal.conan_app import ConanApp
 from conans.model.package_ref import PkgReference
 from conans.model.recipe_ref import RecipeReference
