+++ conan/api/subapi/upload.py
@@ -1,7 +1,7 @@
 from conan.api.model import UploadBundle
 from conan.api.output import ConanOutput
 from conan.api.subapi import api_method
-from conan.api.conan_app import ConanApp
+from conan.internal.conan_app import ConanApp
 from conan.internal.api.select_pattern import SelectPattern
 from conans.client.cmd.uploader import IntegrityChecker, PackagePreparator, UploadExecutor, \
     UploadUpstreamChecker
